Move spawning code out of interaction system
