+++ src/systems/interactions.js
@@ -1,8 +1,5 @@
 /* global AFRAME Ammo NAF */
-import { waitForEvent } from "../utils/async-utils";
 import { paths } from "./userinput/paths";
-import { addMedia } from "../utils/media-utils";
-import { ObjectContentOrigins } from "../object-types";
 
 const handCollisionTargets = new Map();
 AFRAME.registerComponent("is-hand-collision-target", {
@@ -51,36 +48,6 @@ AFRAME.registerComponent("is-remote-hover-target", {
 AFRAME.registerSystem("interaction", {
   updateCursorIntersection: function(intersection) {
     this.rightRemoteHoverTarget = intersection && findRemoteHoverTarget(intersection.object);
-  },
-
-  async spawnObjectRoutine(state, options, superSpawner) {
-    options.entity.object3D.updateMatrices();
-    options.entity.object3D.matrix.decompose(
-      options.entity.object3D.position,
-      options.entity.object3D.quaternion,
-      options.entity.object3D.scale
-    );
-    const data = superSpawner.data;
-    const entity = addMedia(data.src, data.template, ObjectContentOrigins.SPAWNER, data.resolve, data.resize, false)
-      .entity;
-    entity.object3D.position.copy(data.useCustomSpawnPosition ? data.spawnPosition : superSpawner.el.object3D.position);
-    entity.object3D.rotation.copy(data.useCustomSpawnRotation ? data.spawnRotation : superSpawner.el.object3D.rotation);
-    entity.object3D.scale.copy(data.useCustomSpawnScale ? data.spawnScale : superSpawner.el.object3D.scale);
-    entity.object3D.matrixNeedsUpdate = true;
-    state.held = entity;
-
-    superSpawner.activateCooldown();
-    state.spawning = true;
-    // WARNING: waitForEvent is semantically different than entity.addEventListener("body-loaded", ...)
-    // and adding a callback fn via addEventListener will not work unless the callback function
-    // wraps its code in setTimeout(()=>{...}, 0)
-    await waitForEvent("body-loaded", entity);
-    state.spawning = false;
-    entity.object3D.position.copy(data.useCustomSpawnPosition ? data.spawnPosition : superSpawner.el.object3D.position);
-    if (data.centerSpawnedObject) {
-      entity.body.position.copy(options.entity.object3D.position);
-    }
-    entity.object3D.matrixNeedsUpdate = true;
   },
 
   init: function() {
@@ -147,7 +114,6 @@ AFRAME.registerSystem("interaction", {
           const offersConstraint =
             state.hovered.components.tags && state.hovered.components.tags.data[options.constraintTag];
           const isHoldableButton = state.hovered.components.tags && state.hovered.components.tags.data.holdableButton;
-          const superSpawner = state.hovered.components["super-spawner"];
           if (offersConstraint || isHoldableButton) {
             if (
               this.el.is("frozen") ||
@@ -156,8 +122,6 @@ AFRAME.registerSystem("interaction", {
             ) {
               state.held = state.hovered;
             }
-          } else if (superSpawner) {
-            this.spawnObjectRoutine(state, options, superSpawner);
           }
         }
       }
